Split up mobile_app tests to match split up component
